### 2024-12-11 3117269
add acknowledgements and fix typo
