### 2024-12-11 852053e
import int32
