### 2024-12-11 83aedab
add acknowledgements and fix typo
