### 2025-02-19 #7347 @drvinceknight
@Shibabrat I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “NumbaCS: A fast Python package for coherent structure analysis”. You can find more information at the top of this Github issue.
