### 2025-03-26 #7948 @editorialbot
[REVIEW]: NumbaCS: A fast Python package for coherent structure analysis
**Submitting author:** @alb3rtjarvis (<a href="http://orcid.org/0009-0006-8770-6728">Albert Jarvis</a>)
**Repository:** https://github.com/alb3rtjarvis/numbacs
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.1.2
