### 2025-04-28 #7948 @janfb
## NumbaCS Review

Hi @albertjarvis! I've had the chance to review your NumbaCS package for JOSS and am really impressed with what you've built. This looks like a fantastic addition to the coherent structures software landscape, bringing together user-friendliness and computational speed in a way that's been needed. I particularly appreciate how you've designed it to be extensible for future methods.

A few points I'd like to see addressed before final acceptance:

1. **Contributions**: Most GitHub commits seem to be from the first author. Could you clarify Shane Ross's contributions in the paper? JOSS requires that all authors have made substantial contributions, see [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort).

2. **Performance comparisons**: While I was able to reproduce your timing claims, it would be more meaningful to include some comparative benchmarks with related packages (e.g., Dynlab or LCS Tool) for at least one example to better showcase NumbaCS's efficiency.

3. **Testing improvements**:
   - Please add development dependencies (like `pytest`) to your setup, e.g., using dependencies extras like `dev` in `pyproject.toml`
   - The `test_hyperbolic_lcs` test is failing with a shape mismatch on my system - could you address this?
   - Consider adding GitHub Actions for CI/CD workflows (linting, formatting, testing), e.g., using `ruff` and `pre-commit`.
   - A simple example like https://github.com/probabilists/zuko/blob/master/.github/workflows/ci.yml would work well

4. **Documentation**:
   - Your docs are excellent overall!
   - Consider adding direct links to source code in the API reference if possible

5. **Abstract/Summary**: While generally well-written, the first few sentences might be a bit too technical for a non-specialist audience. Consider making the opening more accessible and high-level before diving into the technical details.

I'm excited to see this package move forward - it seems fill an important need in the field. Let me know if you have any questions about my suggestions!
