### 2025-04-29 alb3rtjarvis/numbacs#1 @slayoo
README: clarify the reason why pip is discouraged (and point to the particular issue in "one of the dependencies" that is the problem)
https://github.com/alb3rtjarvis/numbacs/blob/83aedabdc1cdbd21bbf681eb8ba6be2db9cf4874/README.md?plain=1#L37-L38
