### 2025-04-29 alb3rtjarvis/numbacs#3 @slayoo
CI: setup on-release PyPI package upload action
https://github.com/pypa/gh-action-pypi-publish
