### 2025-04-29 alb3rtjarvis/numbacs#4 @slayoo
pyproject.toml: allow installation with Python 3.12 (Numba supports it now)
https://github.com/alb3rtjarvis/numbacs/blob/83aedabdc1cdbd21bbf681eb8ba6be2db9cf4874/pyproject.toml#L10

(and if there was another reason to restrict to Python < 3.12, please add relevant info, best if with a pointer to a relevant issue on GH)
