### 2025-04-29 alb3rtjarvis/numbacs#5 @slayoo
paper: rephrase "standard Python ODE solvers" to point to SciPy (or other package) - otherwise sounds like Python stdlib would contain ODE solvers
https://github.com/alb3rtjarvis/numbacs/blob/83e4cd66e6e518fa992074c59730b3d7ed960775/paper/paper.md?plain=1#L109-L110
