### 2025-04-29 alb3rtjarvis/numbacs#8 @slayoo
consider setting up on-release permanent archival on Zenodo (auto-generates a citeable DOI for each release)
https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content#issuing-a-persistent-identifier-for-your-repository-with-zenodo
