### 2025-04-29 alb3rtjarvis/numbacs#8 @slayoo
metadata can be placed in a repo-stored JSON file: https://help.zenodo.org/docs/github/describe-software/zenodo-json/
