### 2025-05-02 #7948 @slayoo
Hi @alb3rtjarvis & @drvinceknight,

Here are a few notes gathered while getting acquainted with the package:
- https://github.com/alb3rtjarvis/numbacs/issues/1
- https://github.com/alb3rtjarvis/numbacs/issues/2
- https://github.com/alb3rtjarvis/numbacs/issues/3
- https://github.com/alb3rtjarvis/numbacs/issues/4
- https://github.com/alb3rtjarvis/numbacs/issues/5
- https://github.com/alb3rtjarvis/numbacs/issues/6
- https://github.com/alb3rtjarvis/numbacs/issues/7
- https://github.com/alb3rtjarvis/numbacs/issues/8
- https://github.com/alb3rtjarvis/numbacs/pull/9
- https://github.com/alb3rtjarvis/numbacs/issues/10

Concerning the "Contribution and authorship" point in the checklist, leaving unmarked as all commits registered in the repository originate from a single author.

Automating test and example code execution, code coverage analysis, and release uploads using GitHub Actions would be worth setting up for the project.

HTH,
Sylwester
