### 2025-06-06 f7eb359
update pyproject.toml for linting, formatting, coverage
