+++ .github/workflows/ci.yml
@@ -1,0 +1,63 @@
+name: Continuous integration
+
+on:
+  push:
+    branches: [main]
+  pull_request:
+    branches: [main]
+  workflow_dispatch:
+
+jobs:
+  lint_and_format:
+    name: Pre-commit Checks
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+
+      - name: Run pre-commit hooks
+        uses: pre-commit/action@v3.0.1
+        with:
+          extra_args: --all-files --show-diff-on-failure --hook-stage manual
+
+
+  test_and_coverage:
+    name: Test (Python ${{ matrix.python-version }}, OS ${{ matrix.os }})
+    needs: lint_and_format
+    runs-on: ${{ matrix.os }} # Example: Add OS to matrix if needed
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: ['3.10', '3.11']
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Micromamba
+        uses: mamba-org/setup-micromamba@v2
+        with:
+          environment-file: environment-ci.yml
+          environment-name: numbacs-ci-env
+          create-args: >-
+            python=${{ matrix.python-version }} # Sets/overrides Python in the env
+          cache-environment: true
+
+      - name: Install numbacs from source (editable)
+        shell: bash -el {0}
+        run: |
+          python -m pip install --no-deps -e .
+
+      - name: Run Pytest with Coverage
+        shell: bash -el {0}
+        run: pytest --cov=src/numbacs --cov-report=xml --cov-report=term-missing tests/
+
+      - name: Upload coverage to Codecov
+        if: matrix.os == 'ubuntu-latest' && matrix.python-version == '3.11' # Upload from one specific job
+        uses: codecov/codecov-action@v4
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          fail_ci_if_error: true
