+++ environment-dev.yml
@@ -1,0 +1,21 @@
+# environment-dev.yml
+name: numbacs-dev-env
+channels:
+  - conda-forge
+dependencies:
+  - python>=3.10,<3.12
+  - numpy
+  - scipy
+  - numba
+  - interpolation>=2.2.6
+  - numbalsoda
+  - contourpy
+  - pip
+  - pip:
+    - pytest
+    - pytest-cov
+    - ruff
+    - pre-commit
+    - sphinx
+    - sphinx-rtd-theme
+  	
