### 2025-06-06 97394c0
update ci.yml
