+++ .github/workflows/ci.yml
@@ -69,3 +69,4 @@ jobs:
         env:
           NUMBA_NUM_THREADS: 1
         run: python tests/windows_integration_check.py
+
