### 2025-06-06 5dd128f
ruff lint and update pyproject.toml
