+++ examples/hyp_lcs/plot_dg_hyp_lcs.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """
 Double gyre Hyperbolic LCS
 ==========================
@@ -79,8 +78,8 @@ lambda_avg_min = 600
 arclen_flag=True
 
 # extract hyperbolic lcs
-lcs = hyperbolic_lcs(eigval_max, eigvecs, x, y, step_size, steps, lf, lmin, r, nmax, 
-                     dist_tol=dtol, 
+lcs = hyperbolic_lcs(eigval_max, eigvecs, x, y, step_size, steps, lf, lmin, r, nmax,
+                     dist_tol=dtol,
                      nlines=nlines,
                      ep_dist_tol=ep_dist_tol,
                      percentile=percentile,
