+++ examples/hyp_oecs/plot_bickley_ile.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """
 
 Bickley jet iLE
