+++ src/numbacs/integration.py
@@ -37,7 +37,7 @@ def flowmap(funcptr,t0,T,pts,params,method='dop853',rtol=1e-6,atol=1e-8):
         array containing final position of particles pts after integration from [t0,t0+T].
 
     """
-    	
+
     flowmap = np.zeros(pts.shape,float64)
     t_eval = params[0]*np.linspace(t0,t0+T,2)
     if method.lower() == 'dop853':
@@ -50,7 +50,7 @@ def flowmap(funcptr,t0,T,pts,params,method='dop853',rtol=1e-6,atol=1e-8):
             flowmap_tmp,success = lsoda(funcptr, np.array([pts[i,0],pts[i,1]]),
                                          t_eval,rtol=rtol,atol=atol,data=params)
             flowmap[i,:] = flowmap_tmp[-1,:]
-    
+
     return flowmap
 
 
@@ -91,7 +91,7 @@ def flowmap_n(funcptr,t0,T,pts,params,method='dop853',n=2,rtol=1e-6,atol=1e-8):
         array containing times the flowmap is returned at.
 
     """
-    	
+
     npts = len(pts)
     flowmap = np.zeros((npts,n,2),float64)
     t_eval = params[0]*np.linspace(t0,t0+T,n)
@@ -105,7 +105,7 @@ def flowmap_n(funcptr,t0,T,pts,params,method='dop853',n=2,rtol=1e-6,atol=1e-8):
             flowmap_tmp,success = lsoda(funcptr, np.array([pts[i,0],pts[i,1]]),
                                          t_eval,rtol=rtol,atol=atol,data=params)
             flowmap[i,:,:] = flowmap_tmp
-    
+
     return flowmap, params[0]*t_eval
 
 
@@ -160,7 +160,7 @@ def flowmap_grid_2D(funcptr,t0,T,x,y,params,method='dop853',rtol=1e-6,atol=1e-8)
                 flowmap_tmp,success = lsoda(funcptr, np.array([x[i],y[j]]),
                                             t_eval,rtol=rtol,atol=atol,data=params)
                 flowmap[i,j,:] = flowmap_tmp[-1,:]
-    
+
     return flowmap
 
 
@@ -213,7 +213,7 @@ def flowmap_grid_ND(funcptr,t0,T,IC_flat,ndims,params,method='dop853',rtol=1e-6,
             flowmap_tmp,success = lsoda(funcptr, IC_flat[k*ndims:(k+1)*ndims],
                                         t_eval,rtol=rtol,atol=atol,data=params)
             flowmap[k,:] = flowmap_tmp[-1,:]
-    
+
     return flowmap
 
 
@@ -276,7 +276,7 @@ def flowmap_aux_grid_2D(funcptr,t0,T,x,y,params,h=1e-5,eig_main=True,compute_edg
                 for i in prange(nx):
                     if i == 0 or i == nx-1:
                         for j in range(ny):
-                            flowmap_tmp,success = dop853(funcptr, 
+                            flowmap_tmp,success = dop853(funcptr,
                                                          np.array([x[i], y[j]])+
                                                          aux_grid[-1,:],t_eval,
                                                           rtol=rtol,atol=atol,data=params)
@@ -284,14 +284,14 @@ def flowmap_aux_grid_2D(funcptr,t0,T,x,y,params,h=1e-5,eig_main=True,compute_edg
                     else:
                         for j in range(ny):
                              if j == 0 or j == ny-1:
-                                 flowmap_tmp,success = dop853(funcptr, 
+                                 flowmap_tmp,success = dop853(funcptr,
                                                               np.array([x[i], y[j]])+
                                                               aux_grid[-1,:],t_eval,
                                                               rtol=rtol,atol=atol,data=params)
[…]
