### 2025-06-06 71b1c80
ruff format and update pyproject.toml
