+++ src/numbacs/utils.py
@@ -4,8 +4,9 @@ import numba
 from math import floor, pi, cos, sin
 from scipy.interpolate import splprep, splev
 
+
 @njit
-def unravel_index(index,shape):
+def unravel_index(index, shape):
     """
     Numba currently does not implement np.unravel_index so we create an implementation
     here for the specific case where order='C' (row-major).
@@ -25,10 +26,10 @@ def unravel_index(index,shape):
     """
 
     shape = np.flip(shape)
-    arr_ind = np.zeros(len(shape),numba.int64)
+    arr_ind = np.zeros(len(shape), numba.int64)
     ind = index
-    for i in range(len(shape)-1):
-        div_mod = divmod(ind,shape[i])
+    for i in range(len(shape) - 1):
+        div_mod = divmod(ind, shape[i])
         arr_ind[i] = div_mod[1]
         ind = div_mod[0]
 
@@ -38,7 +39,7 @@ def unravel_index(index,shape):
 
 
 @njit
-def ravel_index(inds,shape):
+def ravel_index(inds, shape):
     """
     Finds raveled index corresponding to grid index given by inds from array with
     shape=shape where shape must be a np.ndarray.
@@ -58,14 +59,14 @@ def ravel_index(inds,shape):
     """
 
     r_ind = inds[-1]
-    for i,ind in enumerate(inds[:-1]):
-        r_ind+=ind*np.prod(shape[i+1:])
+    for i, ind in enumerate(inds[:-1]):
+        r_ind += ind * np.prod(shape[i + 1 :])
 
     return r_ind
 
 
 @njit
-def finite_diff_2D(f,i,j,h,axis,direction='c'):
+def finite_diff_2D(f, i, j, h, axis, direction="c"):
     """
     Compute 2nd order partial finite difference in the array f @ [i,j] along axis=axis
     and using a directional difference scheme defined by direction.
@@ -93,30 +94,30 @@ def finite_diff_2D(f,i,j,h,axis,direction='c'):
 
     """
 
-    if axis==0:
-        if direction == 'c':
-            df = (f[i+1,j] - f[i-1,j])/(2*h)
-        elif direction == 'b':
-            df = (3*f[i,j] - 4*f[i-1,j] + f[i-2,j])/(2*h)
-        elif direction == 'f':
-            df = (-f[i+2,j] + 4*f[i+1,j] - 3*f[i,j])/(2*h)
+    if axis == 0:
+        if direction == "c":
+            df = (f[i + 1, j] - f[i - 1, j]) / (2 * h)
+        elif direction == "b":
+            df = (3 * f[i, j] - 4 * f[i - 1, j] + f[i - 2, j]) / (2 * h)
+        elif direction == "f":
+            df = (-f[i + 2, j] + 4 * f[i + 1, j] - 3 * f[i, j]) / (2 * h)
         else:
-            print("Valid difference directions are 'c', 'b', and 'f'" )
-    if axis==1:
-        if direction == 'c':
-            df = (f[i,j+1] - f[i,j-1])/(2*h)
-        elif direction == 'b':
-            df = (3*f[i,j] - 4*f[i,j-1] + f[i,j-2])/(2*h)
-        elif direction == 'f':
-            df = (-f[i,j+2] + 4*f[i,j+1] - 3*f[i,j])/(2*h)
+            print("Valid difference directions are 'c', 'b', and 'f'")
+    if axis == 1:
+        if direction == "c":
+            df = (f[i, j + 1] - f[i, j - 1]) / (2 * h)
[…]
