### 2025-06-10 f8292d6
set config files and run pre-commit
