+++ docs/source/userguide.rst
@@ -8,7 +8,7 @@ started with NumbaCS, a user may be better served by picking an example from the
 :ref:`auto_examples/index:examples gallery` to go through and refering to this section as
 questions arise. An overview of the theory behind the methods presented here is
 covered in the :ref:`theory:theory and implementation` section.
-   
+
 
 Lagrangian
 ----------
@@ -26,8 +26,8 @@ Flows
 To begin using NumbaCS for Lagrangian methods, the user needs to define a flow
 from which coherent structures will be obtained. This is all done through the
 :mod:`numbacs.flows` module. NumbaCS provides predefined flows that the user
-can retrieve and a simple interface for generating the interpolants needed 
-for numerical velocity fields. 
+can retrieve and a simple interface for generating the interpolants needed
+for numerical velocity fields.
 
 
 Predefined flows
@@ -44,24 +44,24 @@ and "abc". To retrieve the "double_gyre" flow,
    from math import copysign
    import numpy as np
    from numbacs.flows import get_predefined_flow
-   
+
    # define integration time
    T = -10.0
    int_direction = copysign(1, T)
-   
+
    # get function pointer for 'double_gyre'
    funcptr, params, domain = get_predefined_flow('double_gyre',
    						  int_direction = int_direction
    						  return_default_params = True,
    						  return_domain = True
    						  parameter_description = True)
-   						  
+
    # set up grid with nx points in x-direction and ny in y-direction
    nx,ny = 401,201
    x = np.linspace(domain[0][0], domain[0][1], nx)
    y = np.linspace(domain[1][0], domain[1][1], ny)
 
-Predefined flows have parameters in their definition that can be modified. A 
+Predefined flows have parameters in their definition that can be modified. A
 default set of parameters can be returned by setting the keyword argument
 ``return_default_params = True`` (it is True by default). In addition,
 by setting ``parameter_description = True`` (it is False by default), a
@@ -74,17 +74,17 @@ coincides with in the definition of the ODE for each flow. For reference, the OD
 
 .. math::
    :label: dg
-   
+
    \dfrac{dx}{dt} &= -\pi A \sin(\pi f)\cos(\pi y) - \alpha x + \eta\\
    \dfrac{dy}{dt} &= \pi A \cos(\pi f)\sin(\pi y) \dfrac{df}{dx} - \alpha y + \eta\\
    f(x,t) &= \epsilon \sin(\omega t + \psi)x^2 + (1 - 2\epsilon \sin(\omega t + \psi)x).
-   
-   
+
+
 **Bickely Jet**
 
 .. math::
    :label: bickley
-   
+
    \psi(x,y,t) = U_0 L(-\tanh(y/L) + \text{sech}^2(y/L&)(A_1 \cos(k_1(x - c_1 t)\\
                                                        &+ A_2 \cos(k_2(x - c_2 t)\\
                                                        &+ A_3 \cos(k_3(x - c_3 t))
@@ -94,10 +94,10 @@ coincides with in the definition of the ODE for each flow. For reference, the OD
 
 .. math::
    :label: `abc`
-   
+
    \dfrac{dx}{dt} &= (A + \alpha\sin(t))\sin(z) + C\cos(y)\\
    \dfrac{dy}{dt} &= B\sin(x) + (A + \alpha\sin(t))\cos(z)\\
-   \dfrac{dz}{dt} &= C\sin(y) + B\cos(x)                                              
+   \dfrac{dz}{dt} &= C\sin(y) + B\cos(x)
 
 
 Numerical flows
@@ -115,7 +115,7 @@ from that function into :func:`numbacs.flows.get_flow_2D`.
    from math import copysign
[…]
