+++ docs/source/index.rst
@@ -9,7 +9,7 @@ Contents
 
 .. toctree::
    :maxdepth: 1
-   
+
    self
    userguide
    theory
@@ -18,10 +18,9 @@ Contents
    releases
 
 .. include:: ../../README.md
-   :parser: myst_parser.sphinx_    
+   :parser: myst_parser.sphinx_
 
 Indices and tables
 ==================
 
 * :ref:`genindex`
-

[…]
