### 2025-06-10 3f0a88a
fix ci.yml shell problem
