### 2025-06-10 517b8e5
fix ci.yml shell problem
