+++ .github/workflows/ci.yml
@@ -19,13 +19,16 @@ jobs:
       - name: Run pre-commit hooks
         uses: pre-commit/action@v3.0.1
         with:
-          extra_args: --all-files --show-diff-on-failure --hook-stage manual
+          extra_args: --all-files --show-diff-on-failure
 
 
   test_and_coverage:
     name: Test (Python ${{ matrix.python-version }}, OS ${{ matrix.os }})
     needs: pre_commit_checks
     runs-on: ${{ matrix.os }}
+    defaults:
+      run:
+        shell: ${{ matrix.os == 'Windows' && 'powershell' || 'bash -el {0}' }}
     strategy:
       fail-fast: false
       matrix:
@@ -42,15 +45,13 @@ jobs:
           environment-name: numbacs-dev-env
           create-args: python=${{ matrix.python-version }}
           cache-environment: true
-          init-shell: bash
+          init-shell: bash powershell
 
       - name: Install numbacs from source (editable)
-        shell: bash -el {0}
         run: |
           python -m pip install --no-deps -e .
 
       - name: Run Pytest with Coverage
-        shell: bash -el {0}
         run: pytest --cov=src/numbacs --cov-report=xml --cov-report=term-missing tests/
 
 #      - name: Upload coverage to Codecov
