### 2025-06-10 9b1d6f5
fix ci.yml shell problem
