### 2025-06-10 29c2072
fix ci: debug ubuntu failing test and windows fail
