### 2025-06-10 0a4987f
fix ci: debug ubuntu failing test
