### 2025-06-10 23c811a
update test_diagnostics.py to compare eigenvector directions
