+++ .github/workflows/ci.yml
@@ -33,8 +33,9 @@ jobs:
       fail-fast: false
       matrix:
         #os: [ubuntu-latest, macos-latest, windows-latest]
-        os: [ubuntu-latest, macos-latest]
-        python-version: ['3.10', '3.11']
+        os: [ubuntu-latest]
+        #python-version: ['3.10', '3.11']
+        python-version: ['3.11']
         exclude:
           - os: windows-latest
             python-version: '3.11'
