### 2025-06-10 f5b95f1
debug ubuntu fail
