+++ .github/workflows/ci.yml
@@ -57,7 +57,8 @@ jobs:
           python -m pip install --no-deps -e .
 
       - name: Run Pytest with Coverage
-        run: pytest -v --cov=src/numbacs --cov-report=xml --cov-report=term-missing tests/
+        run: pytest -v tests/test_diagnostics.py::test_S_eig_2D_func
+        #run: pytest -v --cov=src/numbacs --cov-report=xml --cov-report=term-missing tests/
 
 #      - name: Upload coverage to Codecov
 #        if: matrix.os == 'ubuntu-latest' && matrix.python-version == '3.11' # Upload from one specific job
