### 2025-06-10 bbe8c3e
debug ubuntu fail
