### 2025-06-10 868aec6
debug ubuntu fail
