+++ environment-dev.yml
@@ -3,7 +3,8 @@ channels:
   - conda-forge
 dependencies:
   - python>=3.10,<3.12
-  - numpy
+  - numpy * openblas
+  - libopenblas
   - scipy
   - numba
   - interpolation>=2.2.6
