+++ environment-dev.yml
@@ -4,7 +4,7 @@ channels:
 dependencies:
   - python>=3.10,<3.12
   - numpy
-  - libblas=*=*openblas
+  - nomkl
   - scipy
   - numba
   - interpolation>=2.2.6
