### 2025-06-10 7962db0
debug ubuntu fail
