### 2025-06-10 7e5d540
debug ubuntu fail
