### 2025-06-10 eaea81a
debug ubuntu fail
