### 2025-06-10 3aa538e
update tests and test full ubuntu CI
