### 2025-06-11 cd0ee23
clean up test_diagnostics.py
