+++ .github/workflows/ci.yml
@@ -33,9 +33,9 @@ jobs:
       fail-fast: false
       matrix:
         #os: [ubuntu-latest, macos-latest, windows-latest]
-        os: [ubuntu-latest]
+        os: [windows-latest]
         #python-version: ['3.10', '3.11']
-        python-version: ['3.11']
+        python-version: ['3.10']
         exclude:
           - os: windows-latest
             python-version: '3.11'
@@ -56,9 +56,13 @@ jobs:
         run: |
           python -m pip install --no-deps -e .
 
-      - name: Run Pytest with Coverage
-        #run: pytest -v tests/test_diagnostics.py::test_S_eig_2D_func
-        run: pytest -v --cov=src/numbacs --cov-report=xml --cov-report=term-missing tests/
+      - name: Run Windows Debug Script
+        if: runner.os == 'Windows'
+        run: python windows_debug_script.py
+
+#      - name: Run Pytest with Coverage
+#        #run: pytest -v tests/test_diagnostics.py::test_S_eig_2D_func
+#        run: pytest -v --cov=src/numbacs --cov-report=xml --cov-report=term-missing tests/
 
 #      - name: Upload coverage to Codecov
 #        if: matrix.os == 'ubuntu-latest' && matrix.python-version == '3.11' # Upload from one specific job
