### 2025-06-11 fa91939
run windows CI tests
