+++ .github/workflows/ci.yml
@@ -56,9 +56,12 @@ jobs:
         run: |
           python -m pip install --no-deps -e .
 
-      - name: Run Windows Debug Script
-        if: runner.os == 'Windows'
-        run: python windows_debug_script.py
+      - name: Run Pytest with Coverage
+        run: pytest -v --cov=src/numbacs --cov-report=xml --cov-report=term-missing tests/
+
+#      - name: Run Windows Debug Script
+#        if: runner.os == 'Windows'
+#        run: python windows_debug_script.py
 
 #      - name: Run Pytest with Coverage
 #        #run: pytest -v tests/test_diagnostics.py::test_S_eig_2D_func
