+++ environment-dev.yml
@@ -4,7 +4,6 @@ channels:
 dependencies:
   - python>=3.10,<3.12
   - numpy
-  - nomkl
   - scipy
   - numba
   - interpolation>=2.2.6
