### 2025-06-11 7187d88
run windows CI tests without coverage
