+++ .github/workflows/ci.yml
@@ -57,7 +57,8 @@ jobs:
           python -m pip install --no-deps -e .
 
       - name: Run Pytest with Coverage
-        run: pytest -v --cov=src/numbacs --cov-report=xml --cov-report=term-missing tests/
+        run: pytest -v --no-cov tests/
+#        run: pytest -v --cov=src/numbacs --cov-report=xml --cov-report=term-missing tests/
 
 #      - name: Run Windows Debug Script
 #        if: runner.os == 'Windows'
