### 2025-06-11 f802d32
run windows CI tests without coverage
