+++ .github/workflows/ci.yml
@@ -56,8 +56,8 @@ jobs:
         run: |
           python -m pip install --no-deps -e .
 
-      - name: Run Pytest with Coverage
-        run: pytest -v --no-cov tests/
+      - name: Run Pytest without Coverage
+        run: pytest -v tests/test_integration.py tests/test_diagnostics.py
 #        run: pytest -v --cov=src/numbacs --cov-report=xml --cov-report=term-missing tests/
 
 #      - name: Run Windows Debug Script
