### 2025-06-11 83b3ced
run windows CI tests without coverage
