### 2025-06-11 ac8caf0
add light windows integration functionality test
