+++ environment-ci.yml
@@ -1,0 +1,17 @@
+name: numbacs-ci
+channels:
+  - conda-forge
+dependencies:
+  - python>=3.10,<3.12
+  - numpy
+  - scipy
+  - numba
+  - interpolation>=2.2.6
+  - numbalsoda
+  - contourpy
+  - pip
+  - pip:
+    - pytest
+    - pytest-cov
+    - ruff
+    - pre-commit
