### 2025-06-11 3cef621
update workflow for windows tests
