+++ tests/windows_integration_check.py
@@ -1,3 +1,4 @@
+print("Importing basic packages...")
 import numpy as np
 import sys
 
