+++ .github/workflows/ci.yml
@@ -60,13 +60,13 @@ jobs:
         if: runner.os != 'Windows'
         run: pytest -v --cov=src/numbacs --cov-report=xml tests/
 
-#      - name: Run Pytest excluding integration (Windows)
-#        if: runner.os == 'Windows'
-#        run: pytest -v --cov=src/numbacs --cov-report=xml --ignore=tests/test_integration.py tests/
-
       - name: Run light integration test for Windows
         if: runner.os == 'Windows'
         run: python tests/windows_integration_check.py
+
+#      - name: Run Pytest excluding integration (Windows)
+#        if: runner.os == 'Windows'
+#        run: pytest -v --cov=src/numbacs --cov-report=xml --ignore=tests/test_integration.py tests/
 
 #  windows_integration_check:
 #    name: Windows Integration Check
