+++ windows_integration_check.py
@@ -1,4 +1,3 @@
-print("Importing basic packages...")
 import numpy as np
 import sys
 
