### 2025-06-11 e0d625f
try windows_check_integration with assert
