+++ windows_integration_check.py
@@ -37,6 +37,6 @@ except Exception as e:
     sys.exit(1)
 
 print("4. Loading in test data and comparing result...")
-# fm_expected = np.load("./tests/testing_data/fm_data.npy")
-# assert np.allclose(fm, fm_expected)
+fm_expected = np.load("./tests/testing_data/fm_data.npy")
+assert np.allclose(fm, fm_expected)
 print("--- Script finished successfully ---")
