### 2025-06-11 454b967
try windows_check_integration with assert (correct load)
