### 2025-06-11 dfa50ab
try full testing CI run
