+++ windows_integration_check.py
@@ -36,7 +36,8 @@ except Exception as e:
     print(f"ERROR: The flowmap_grid_2D function failed with a Python exception: {e}")
     sys.exit(1)
 
-print("4. Loading in test data and comparing result...")
+print("4. Loading in expected data and comparing result...")
 fm_expected = np.load("./tests/testing_data/fm.npy")
 assert np.allclose(fm, fm_expected)
+print("   ... SUCCESS: test data and expected data match.")
 print("--- Script finished successfully ---")
