### 2025-06-11 cfacb81
run full CI tests
