+++ .github/workflows/ci.yml
@@ -32,10 +32,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        #os: [ubuntu-latest, macos-latest, windows-latest]
-        os: [windows-latest]
-        #python-version: ['3.10', '3.11']
-        python-version: ['3.10']
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: ['3.10', '3.11']
         exclude:
           - os: windows-latest
             python-version: '3.11'
@@ -56,7 +54,7 @@ jobs:
         run: |
           python -m pip install --no-deps -e .
 
-      - name: Run Pytest and Coverage (UNIX-like)
+      - name: Run Pytest and Coverage (UNIX)
         if: runner.os != 'Windows'
         run: pytest -v --cov=src/numbacs --cov-report=xml tests/
 
@@ -69,41 +67,3 @@ jobs:
         env:
           NUMBA_NUM_THREADS: 1
         run: python windows_integration_check.py
-
-#  windows_integration_check:
-#    name: Windows Integration Check
-#    needs: test_suite
-#    if: success()
-#    runs-on: windows-latest
-
-#    steps:
-#      - uses: actions/checkout@v4
-#      - name: Set up Micromamba
-#        uses: mamba-org/setup-micromamba@v2
-#        with:
-#          environment-file: environment-ci.yml
-#          environment-name: numbacs-ci-env
-#          create-args: python=3.10
-#          cache-environment: true
-#          init-shell: bash powershell
-
-#      - name: Install numbacs
-#        run: python -m pip install --no-deps -e .
-
-#      - name: Run Windows Integration Confidence Check Script
-
-
-#      - name: Run Windows Debug Script
-#        if: runner.os == 'Windows'
-#        run: python windows_debug_script.py
-
-#      - name: Run Pytest with Coverage
-#        #run: pytest -v tests/test_diagnostics.py::test_S_eig_2D_func
-#        run: pytest -v --cov=src/numbacs --cov-report=xml --cov-report=term-missing tests/
-
-#      - name: Upload coverage to Codecov
-#        if: matrix.os == 'ubuntu-latest' && matrix.python-version == '3.11' # Upload from one specific job
-#        uses: codecov/codecov-action@v4
-#        with:
-#          token: ${{ secrets.CODECOV_TOKEN }}
-#          fail_ci_if_error: true
