### 2025-06-12 9fa889b
move windows integration test and update CI
