### 2025-06-12 3006aa5
fix broken typo in CI
