### 2025-06-13 41f47f5
remove powershell for CI
