+++ .github/workflows/ci.yml
@@ -28,11 +28,12 @@ jobs:
     runs-on: ${{ matrix.os }}
     defaults:
       run:
-        shell: ${{ matrix.os == 'Windows' && 'powershell' || 'bash -el {0}' }}
+        shell: bash -el {0}
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
+#        os: [ubuntu-latest, macos-latest, windows-latest]
+        os: [windows-latest]
         python-version: ['3.10', '3.11']
         exclude:
           - os: windows-latest
@@ -48,7 +49,7 @@ jobs:
           environment-name: numbacs-ci
           create-args: python=${{ matrix.python-version }}
           cache-environment: true
-          init-shell: bash powershell
+          init-shell: bash
 
       - name: Install numbacs from source
         run: |
