### 2025-06-13 5d36733
update dev env with proper docs deps
