### 2025-06-13 3ff1dac
update with proper docs deps
