### 2025-06-13 73c564c
update CI with new envs
