+++ .github/workflows/ci.yml
@@ -32,8 +32,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-#        os: [ubuntu-latest, macos-latest, windows-latest]
-        os: [windows-latest]
+        os: [ubuntu-latest, macos-latest, windows-latest]
         python-version: ['3.10', '3.11']
         exclude:
           - os: windows-latest
@@ -45,7 +44,7 @@ jobs:
       - name: Set up Micromamba
         uses: mamba-org/setup-micromamba@v2
         with:
-          environment-file: environment-ci.yml
+          environment-file: .github/envs/environment-ci.yml
           environment-name: numbacs-ci
           create-args: python=${{ matrix.python-version }}
           cache-environment: true
@@ -53,7 +52,7 @@ jobs:
 
       - name: Install numbacs from source
         run: |
-          python -m pip install --no-deps -e .
+          python -m pip install --no-deps .
 
       - name: Run Pytest and Coverage (UNIX)
         if: runner.os != 'Windows'
