### 2025-06-13 0c20d3d
explicitly list why conda is strongly recommended, add description for 'similar software'
