### 2025-06-13 52513b2
fix typo and update reason for conda preference
