### 2025-06-13 483bb0e
fix typo and update reason for conda preference
