+++ docs/source/auto_examples/elliptic_oecs/plot_bickley_elliptic_oecs.rst
@@ -158,7 +158,7 @@ Plot the elliptic OECS over the IVD field.
     # sphinx_gallery_thumbnail_number = 1
     fig,ax = plt.subplots(dpi=200)
     ax.contourf(x,y,ivd.T,levels=80)
-    ax.set_aspect('equal')   
+    ax.set_aspect('equal')
     for rcv,c in elcs:
         ax.plot(rcv[:,0],rcv[:,1],lw=1.5)
         ax.scatter(c[0],c[1],1.5)
