### 2025-06-13 4d26b4d
fix typos, add conda preference details, add benchmarks link
