### 2025-06-13 d5d9bfc
update docs and re-run examples
