+++ docs/source/auto_examples/hyp_oecs/plot_qge_hyp_oecs.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """
 Quasi-geostrophic hyperbolic OECS
 =================================
@@ -22,16 +21,16 @@ import matplotlib.pyplot as plt
 # Load velocity data and set up domain.
 
 # load in qge velocity data
-u = np.load('../data/qge/qge_u.npy')
-v = np.load('../data/qge/qge_v.npy')
+u = np.load("../data/qge/qge_u.npy")
+v = np.load("../data/qge/qge_v.npy")
 
 # set up domain
-nt,nx,ny = u.shape
-x = np.linspace(0,1,nx)
-y = np.linspace(0,2,ny)
-t = np.linspace(0,1,nt)
-dx = x[1]-x[0]
-dy = y[1]-y[0]
+nt, nx, ny = u.shape
+x = np.linspace(0, 1, nx)
+y = np.linspace(0, 2, ny)
+t = np.linspace(0, 1, nt)
+dx = x[1] - x[0]
+dy = y[1] - y[0]
 # %%
 # S eigenvalues, eigenvectors
 # ---------------------------
@@ -39,8 +38,8 @@ dy = y[1]-y[0]
 
 k0 = 15
 # compute eigenvalues/vectors of Eulerian rate of strain tensor
-eigvals,eigvecs = S_eig_2D_data(u[k0,:,:],v[k0,:,:],dx,dy)
-s2 = eigvals[:,:,1]
+eigvals, eigvecs = S_eig_2D_data(u[k0, :, :], v[k0, :, :], dx, dy)
+s2 = eigvals[:, :, 1]
 
 # %%
 # Hyperbolic OECS saddles
@@ -52,24 +51,23 @@ r = 0.2
 h = 1e-4
 steps = 4000
 maxlen = 0.05
-minval = np.percentile(s2,50)
+minval = np.percentile(s2, 50)
 n = 10
 
 # compute hyperbolic_oecs
-oecs = hyperbolic_oecs(s2,eigvecs,x,y,r,h,steps,maxlen,minval,n=n)
+oecs = hyperbolic_oecs(s2, eigvecs, x, y, r, h, steps, maxlen, minval, n=n)
 
 # %%
 # Plot all OECS
 # -------------
 # Plot the OECS overlaid on iLE.
-fig,ax = plt.subplots(dpi=200)
-ax.contourf(x,y,s2.T,levels=np.linspace(0,np.percentile(s2,99.5),51),
-            extend='both',zorder=0)
+fig, ax = plt.subplots(dpi=200)
+ax.contourf(x, y, s2.T, levels=np.linspace(0, np.percentile(s2, 99.5), 51), extend="both", zorder=0)
 
 for k in range(len(oecs)):
-    ax.plot(oecs[k][0][:,0],oecs[k][0][:,1],'r',lw=1)
-    ax.plot(oecs[k][1][:,0],oecs[k][1][:,1],'b',lw=1)
-ax.set_aspect('equal')
+    ax.plot(oecs[k][0][:, 0], oecs[k][0][:, 1], "r", lw=1)
+    ax.plot(oecs[k][1][:, 0], oecs[k][1][:, 1], "b", lw=1)
+ax.set_aspect("equal")
 plt.show()
 # %%
 # Advect OECS
@@ -82,41 +80,55 @@ from numbacs.utils import gen_filled_circ
 from numbacs.integration import flowmap_n
 
 # get funcptr, set parameters for integration, and integrate
-grid_vel, C_eval_u, C_eval_v = get_interp_arrays_2D(t,x,y,u,v)
[…]
