### 2025-06-13 8b8d561
update paper to incorporate reviewer feedback
