+++ paper/paper.md
@@ -25,17 +25,20 @@ bibliography: paper.bib
 
 # Summary
 
-`NumbaCS` (Numba Coherent Structures) is a Python package that implements a
-variety of coherent structure methods in an efficient and user-friendly manner.
-"Coherent structure methods" refer to any method which can be used to infer or
-extract Lagrangian and objective Eulerian coherent structures. The theory behind
-these methods have been developed over the last few decades with the aim of
-extending many of the important invariant objects from time-independent
-dynamical systems theory to the more general setting where a system may have
-arbitrary time dependence and may only be known or defined for some finite time.
-These time-dependent systems are ubiquitous in the context of geophysical and
-engineering flows where the evolution of the velocity field depends on time and
-velocity data representing these flows is not available for all time. By
+`NumbaCS` (Numba Coherent Structures) is a Python package that efficiently
+implements a variety of methods for studying material transport in time-dependent
+fluid flows. It leverages `Numba` -- a high performance Python compiler
+for generating optimized machine code from Python functions -- along with other
+`Numba`-compatible packages behind the scenes, producing fast and user-friendly
+implementations of coherent structure methods. "Coherent structure methods" refer
+to any method which can be used to infer or extract Lagrangian and objective
+Eulerian coherent structures. The theory behind these methods have been developed
+over the last few decades with the aim of extending many of the important invariant
+objects from time-independent dynamical systems theory to the more general setting
+where a system may have arbitrary time dependence and may only be known or defined
+for some finite time. These time-dependent systems are ubiquitous in the context of
+geophysical and engineering flows where the evolution of the velocity field depends
+on time and velocity data representing these flows is not available for all time. By
 extending the ideas from the time-independent setting to the more general
 time-dependent setting, important transient objects (coherent structures) can
 be identified which govern how material is transported within a flow.
@@ -47,20 +50,20 @@ persons lost at sea.
 
 # Statement of need
 
-As theory and implementations of coherent structures have been developed [@haller1998; @shadden2005; 
-@mathur2007; @haller2011; @farazmand2012; @schindler2012; @haller2013; 
+As theory and implementations of coherent structures have been developed [@haller1998; @shadden2005;
+@mathur2007; @haller2011; @farazmand2012; @schindler2012; @haller2013;
 @haller2016; @serra2016; @nolan2020] and the utility of these tools
 has been demonstrated over the last two decades [@dutoit2010; @rutherford2012;
-@peacock2013; @serra2017; @liu2018; @nolanpollution2020; @gunther2021; 
+@peacock2013; @serra2017; @liu2018; @nolanpollution2020; @gunther2021;
 @pretorius2023; @curbelo2023], there has been a steadily growing
-interest in using these methods for real-world applications. 
+interest in using these methods for real-world applications.
 Early on, software implementations were largely contained to in-house packages
 developed by applied mathematicians and engineers advancing the theory. Over
 the years, there have been a number of software packages developed in an attempt
 to provide implementations of some of these methods for practitioners outside
 of the field. Some provide a friendly interface for users (`Dynlab` -- @dynlab;
  `LCS MATLAB Kit` -- @lcskit), others aim to provide efficient implementations of
-specific methods (sometimes in specific circumstances) 
+specific methods (sometimes in specific circumstances)
 (`Lagrangian` -- @lagrangian; `Newman` -- @dutoitthesis2010;
[…]
