+++ .github/workflows/ci.yml
@@ -3,6 +3,8 @@ name: Continuous integration
 on:
   push:
     branches: [main]
+  pull_request:
+    branches: [main]
   workflow_dispatch:
 
 jobs:
