### 2025-06-13 alb3rtjarvis/numbacs#1 @alb3rtjarvis
Clarified in the README and linked to relevant issues.
