### 2025-06-13 alb3rtjarvis/numbacs#4 @alb3rtjarvis
The issue is the numbalsoda package which does not support >3.11. I have added a comment to the pyproject.toml. I could add it elsewhere too if you think that would be helpful.
