### 2025-06-13 alb3rtjarvis/numbacs#5 @alb3rtjarvis
I have reworded that section of the paper to reflect these comments.
