### 2025-06-13 alb3rtjarvis/numbacs#10 @alb3rtjarvis
Tests are now set up to run on push and pull requests to main (see [`.github/workflows/ci.yml`](https://github.com/alb3rtjarvis/numbacs/blob/main/.github/workflows/ci.yml))
