### 2025-06-13 #7948 @alb3rtjarvis
Hi @slayoo! Thanks for all the thoughful feedback! I have addressed most of your points on the GitHub issues but I will address them again here for completeness.

- **README why is pip discourged?:** This is due to some reported issues on numbalsoda github page ([#20](https://github.com/Nicholaswogan/numbalsoda/issues/20) and [#29](https://github.com/Nicholaswogan/numbalsoda/issues/29)). These issues seem to mainly affect Windows installations. I have updated the README to reflect this.
- **CI: set up code coverage analysis:** I will just copy my response from the github issue
   * This was a bit tricky. Using NUMBA_DISABLE_JIT=1 threw errors due to numbalsoda failing without JIT (if I am not mistaken, I believe this is due to `@cfunc` decorator being disabled as well which is needed to create the functions for the ODE solvers from numbalsoda to act on). I cannot test any integration functions without numbalsoda properly working. I also tried to get coverage results without disabling JIT as Numba docs imply this is possible (see [code coverage](https://numba.readthedocs.io/en/stable/user/code_coverage.html)) but it was not producing accurate coverage reports. It seems there may still be some bugs to iron out with this (see [#9859](https://github.com/numba/numba/issues/9859)) or perhaps I am missing something about how to get this working. If you have any suggestions I am happy to give them a try, otherwise my plan was to wait for the next Numba release as they seem to be actively working on compiled code coverage and related bugs (see [#9862](https://github.com/numba/numba/pull/9862)).

- **CI: setup on-release PyPI package upload:** I set this up in [`.github/workflows/release.yml`](https://github.com/alb3rtjarvis/numbacs/blob/main/.github/workflows/release.yml) but have not yet tested it as I am not ready for a new full release. If you would like me to test it before moving forward I can set it up to publish on TestPyPI on a pre-release and test it that way.
- **pyproject.toml allow Python 3.12 (Numba now supports it):** The issue is the numbalsoda package which does not support >3.11. I have added a comment to the pyproject.toml. I could add it elsewhere too if you think that would be helpful.
- **paper: rephrase "standard ODE solvers":** I have reworded that section of the paper to reflect this and now explicitly mention SciPy ODE solvers.
- **paper: explain Numba earlier:** I have added a sentence briefly describing Numba in the `Summary` section.
- **paper: replace raster graphics with vector graphics:** In my experience, vector graphics of many-layered filled contour plots are usually time-consuming to render and make the document they are in sluggish. But I am happy to give this a try if you think it is worth it.
- **Consider setting up on-release archival on Zenodo:** This has been set up on Zenodo. Again, if you prefer I test it before moving forward I can do the pre-release as mentioned for PyPI publishing.
- **docs: fix broken Sphinx website:** Thanks for catching this, I have merged your pull request.
- **CI setup test and example code execution on GitHub Actions:** I have set up CI for testing though I was a little unsure about what you meant about setting up Actions for example code execution (there is a great deal of example code in the docs, some with runtimes).
- **Contributing and authorship:** The other reviewer brought up the same point so I will paste my response here
   * NumbaCS was born from discussions Shane Ross and I had about the lack of easy-to-use and efficient coherent structure software. He and I met regularly over the duration of the development of the package, and he contributed to project direction through algorithm design and validation, theoretical discussions about methods, and suggestions for the addition of methods from conversations with practitioners that planned to use the software. He was instrumental in the development of the package.

Thank you for all the feedback and suggestions! They taught me quite a bit about CI and definitely put the package in a better place for future development and collaboration. I know there are a few things I was unable to fully address at the moment so please let me know if you like me to revisit anything before moving forward. Thanks again!
