### 2025-06-13 015ef67
make draft of paper and try to reformat
