### 2025-06-13 c9b9f7c
make draft of paper and try to reformat
