### 2025-06-13 4a049dd
make draft of paper and reformat
