+++ paper/paper.md
@@ -175,6 +175,7 @@ at t~0~ = 0.5. Total runtime per iterate: ~0.0452s
 (IVD: ~0.0002s; elliptic OECS extraction: ~0.045s).
 All are computed over a 257x513 grid.
 
+\newpage
 
 **Analytical Flow (Bickley jet)**
 
@@ -182,7 +183,7 @@ All are computed over a 257x513 grid.
 
 [Bickley jet elliptic LCS](https://numbacs.readthedocs.io/en/latest/auto_examples/elliptic_lcs/plot_bickley_elliptic_lcs.html#sphx-glr-auto-examples-elliptic-lcs-plot-bickley-elliptic-lcs-py)
 at t~0~ = 0, integration time T = 40 days. Total runtime per
-iterate: ~9.200s (flowmap: ~5.050s; LAVD: ~4.140s; elliptic LCS extraction: ~0.010s).
+iterate: ~5.065s (flowmap: ~4.490s; LAVD: ~0.565s; elliptic LCS extraction: ~0.010s).
 Computed over 482x121 grid.
 
 **Numerical Flow (MERRA-2)**
