### 2025-07-07 #7948 @slayoo
@alb3rtjarvis, thanks for all the work. kudos.

The single most important (and very "principal" given JOSS reproducibility and code curation principles) thing, I would recommend to address, is the persistent code archival at Zenodo. If, effectively, testing the PyPI upload workflow for `test.pypi.org` could be part of it, even better!

Note that according to JOSS workflow (https://joss.readthedocs.io/en/latest/submitting.html#review-process):
> ... authors will make a tagged release of the software, and deposit a copy of the repository with a data-archiving service such as [Zenodo](https://zenodo.org/) or [figshare](https://figshare.com/), get a DOI for the archive ...
