+++ src/numbacs/__init__.py
@@ -3,4 +3,4 @@ Top-level package for numbacs.
 """
 
 __author__ = """Albert Jarvis"""
-__version__ = "0.1.1"
+__version__ = "0.1.2rc1"
