+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "hatchling.build"
 [project]
 name = "numbacs"
 dynamic = ["version"]
-description = 'Numba accelerated coherent structure package.'
+description = 'Numba-accelerated coherent structure package.'
 readme = "README.md"
 requires-python = ">=3.10, <3.12"    # numbalsoda currently does not support >3.11
 license = "MPL-2.0"
