### 2025-07-08 674b707
update version to 0.1.2rc2
