### 2025-07-08 #7948 @alb3rtjarvis
Hi @slayoo, I have tested the workflow for Zenodo using their sandbox: [https://sandbox.zenodo.org/records/286746](https://sandbox.zenodo.org/records/286746) and also published to `test.pypi.org` using the workflow: [https://test.pypi.org/project/numbacs/0.1.2rc4/](https://test.pypi.org/project/numbacs/0.1.2rc4/). I have switched off the sandbox archiving and switched on the production Zenodo archiving for the full tagged release upon competition of the review. Thanks for this suggestion, there were a few minor errors in the Zenodo json causing issues.

Thanks again for all the feedback and great suggestions!
