### 2025-07-15 f7f171e
add examples gallery link to readme
