### 2025-07-29 #7948 @slayoo
I confirm (just accessed NumbaCS through the Zenodo sandbox). Thanks!
