### 2025-08-14 #7948 @alb3rtjarvis
@drvinceknight I apologize for the delayed answer, for some reason I did not see these recent comments. I have released a new tagged version and archive, and have updated the title to match the paper title.

Version: 0.1.2
DOI: https://doi.org/10.5281/zenodo.16877771

Please let me know if you need anything else from me.
