### 2025-08-16 27cd13f
Update version number for JOSS release
