### 2025-08-16 a52e0bb
update hatching build version to fix metadata-version error
