### 2025-09-05 #7948 @alb3rtjarvis
Hi @drvinceknight, upon reviewing this, it appears there was an inconsistency with author affiliations between the paper and the zenodo archive (on zenodo both authors were listed as affiliated with "Virginia Tech", while the paper lists each as separate affiliations based on the departments). I have updated the zenodo archive to match the paper. Could this have been causing the error?
