### 2025-09-15 #7948 @drvinceknight
@alb3rtjarvis it looks like you have a duplicate entry in your bibtex file for `serra2017`: [line 443](https://github.com/alb3rtjarvis/numbacs/blob/paper/paper/paper.bib#L443) and [line 321](https://github.com/alb3rtjarvis/numbacs/blob/paper/paper/paper.bib#L321).
