### 2025-09-15 59b43f9
remove duplicate bib entry
