+++ paper/paper.bib
@@ -318,18 +318,6 @@
   pages={R4}
 }
 
-@article{serra2017,
-  author = {Serra, Mattia  and Haller, George },
-  title = "{Efficient computation of null geodesics with applications to coherent vortex detection}",
-  journal = {Proceedings of the Royal Society A: Mathematical, Physical and Engineering Sciences},
-  volume = {473},
-  number = {2199},
-  pages = {20160807},
-  year = {2017},
-  doi = {10.1098/rspa.2016.0807},
-  abstract = { Recent results suggest that boundaries of coherent fluid vortices (elliptic coherent structures) can be identified as closed null geodesics of appropriate Lorentzian metrics defined on the flow domain. Here we derive an automated method for computing such null geodesics based on the geometry of the underlying geodesic flow. Our approach simplifies and improves existing procedures for computing variationally defined Eulerian and Lagrangian vortex boundaries. As an illustration, we compute objective vortex boundaries from satellite-inferred ocean velocity data. A MATLAB implementation of our method is available at https://github.com/MattiaSerra/Closed-Null-Geodesics-2D. }
-}
-
 @article{haller2016,
   title="{Defining coherent vortices objectively from the vorticity}",
   volume={795},
