### 2025-09-15 #7948 @alb3rtjarvis
Proof looks good to me!
