### 2025-09-15 #7948 @danielskatz
@alb3rtjarvis - As track editor, I've suggested a number of small changes in https://github.com/alb3rtjarvis/numbacs/pull/11 - please merge this, or let me know what you disagree with, then we can continue.
