### 2025-09-15 #7948 @danielskatz
@alb3rtjarvis - can you again check this version and let me know if you think it looks ok? I'll do the same.
