### 2025-09-15 #7948 @danielskatz
Congratulations to @alb3rtjarvis (Albert Jarvis) and co-author on your JOSS publication!!

JOSS depends fully on volunteers, and we couldn't do this without you
