### 2025-09-15 #7948 @alb3rtjarvis
Thanks @drvinceknight and @danielskatz for your work as editors! And a big thanks for all the thoughtful feedback and suggestions @janfb and @slayoo, it was greatly appreciated!

I have checked the pull request and the DOI resolves so everything looks good to me.
